Add explicit int precision for buggy gles implementations

+++ shaders/line.vert
@@ -20,7 +20,7 @@ uniform vec2 scale;
 uniform highp vec2 start;
 uniform highp vec2 end;
 uniform float width;
-uniform int cap;
+uniform mediump int cap;
 
 uniform vec4 startColor;
 uniform vec4 endColor;
